moved listingitem blackbox tests back to the blackbox test folder root
